+++ CHANGELOG.md
@@ -23,6 +23,7 @@ especially the [Rust flavour](https://doc.rust-lang.org/cargo/reference/semver.h
 
 ### Changed
 - renamed `ConstBehaviorTreeElementList` to `BehaviorTreeElementList`
+- reworked error handling
 
 ### Fixed
 - behavior `Precondition`s else branch value is now case independant